Added some code refactoring
