--- lesson2/helpers/cardsHelpers.js
@@ -1,14 +1,0 @@
-module.exports = {
-  getCardBJValue: value => {
-    if(isNaN(value)) {
-      if(value === 'A') {
-        value = 11
-      }
-      else {
-        value = 10
-      }
-    }
-
-    return value
-  }
-}
