+++ lesson2/helpers/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  isAnswerYes: answer => answer === 'y' || answer === 'Y',
+  isAnswerNo: answer => answer === 'n' || answer === 'N'
+}
