+++ lesson2/lib/bj/views/index.js
