+++ lesson2/package.json
@@ -4,6 +4,7 @@
   "description": "GB NodeJS course homeworks",
   "main": "index.js",
   "scripts": {
+    "start": "node index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "MaratOM",
